added sign up result confirmation
