+++ client_app/src/components/volunteer/ShelterInfo.js
@@ -1,0 +1,12 @@
+import { Address } from './Address';
+
+export const ShelterInfo = ({ shelter, showLocation = true }) => (
+  <>
+    <div className="shelter-name">{shelter?.name}</div>
+    {showLocation && (
+      <div className="shelter-location">
+        <Address address={shelter.address}/>
+      </div>
+    )}
+  </>
+);
